requirements.txt
# Core ML / NLP
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
faiss-cpu>=1.8.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Utilities
pandas>=2.2.0
